+++ husky_components/toolbar/main.js
@@ -175,12 +175,22 @@ define(function() {
          * event to change a buttons selected dropdown-item
          *
          * @event husky.toolbar.[INSTANCE_NAME.]item.change
-         * @param {string} button The id of the button
-         * @param {string} item the id or the index of the dropdown-item
-         * @param {boolean} executeCallback if true callback of dropdown item gets executed
+         * @param {string} buttonId The id of the button
+         * @param {string} itemId The id or the index of the dropdown-item
+         * @param {boolean} executeCallback If true callback of dropdown item gets executed
          */
         ITEM_CHANGE = function() {
             return createEventName.call(this, 'item.change');
+        },
+
+        /**
+         * Event to reset a buttons and it's dropdown-items into original state.
+         *
+         * @event husky.toolbar.[INSTANCE_NAME.]item.reset
+         * @param {string} buttonId The id of the button
+         */
+        ITEM_RESET = function() {
+            return createEventName.call(this, 'item.reset');
         },
 
         /**
@@ -330,6 +340,8 @@ define(function() {
 
             this.sandbox.on(ITEM_CHANGE.call(this), changeButton.bind(this));
 
+            this.sandbox.on(ITEM_RESET.call(this), resetButton.bind(this));
+
             this.sandbox.on(BUTTON_SET.call(this), function(button, newData) {
                 changeMainListItem.call(this, this.items[button].$el, newData);
             }.bind(this));
@@ -356,29 +368,50 @@ define(function() {
          * Changes a button text and icon.
          *
          * @param {String} button
-         * @param {String|Number} id
+         * @param {String|Number} itemId
          * @param {Bool} executeCallback
          */
-        changeButton = function(button, id, executeCallback) {
-            if (!!this.items[button]) {
-                this.items[button].initialized.then(function() {
-                    // if id is null - unset icon and title when no item selected
-                    if (id === null) {
-                        resetMainListItem.call(this, this.items[button].$el);
-                        return;
-                    }
-
-                    // update icon
-                    var index = getItemIndexById.call(this, id, this.items[button]);
-                    changeMainListItem.call(this, this.items[button].$el, this.items[button].dropdownItems[index]);
-                    this.sandbox.emit(ITEM_MARK.call(this), this.items[button].dropdownItems[index].id);
-                    if (executeCallback === true || !!this.items[button].dropdownItems[index].callback) {
-                        if (typeof this.items[button].dropdownItems[index].callback === 'function') {
-                            this.items[button].dropdownItems[index].callback();
-                        }
-                    }
-                }.bind(this));
+        changeButton = function(buttonId, itemId, executeCallback) {
+            // check if button exists
+            if (!this.items[buttonId]) {
+                return;
             }
+
+            var button = this.items[buttonId];
+
+            button.initialized.then(function() {
+                // if id is null - unset icon and title when no item selected
+                if (itemId === null) {
+                    resetMainListItem.call(this, button.$el);
+                    return;
+                }
+
+                // update icon
+                var index = getItemIndexById.call(this, itemId, button);
+                changeMainListItem.call(this, button.$el, button.dropdownItems[index]);
+                this.sandbox.emit(ITEM_MARK.call(this), button.dropdownItems[index].id);
+                if (executeCallback === true || !!button.dropdownItems[index].callback) {
+                    if (typeof button.dropdownItems[index].callback === 'function') {
+                        button.dropdownItems[index].callback();
+                    }
+                }
+            }.bind(this));
+        },
+
+        /**
+         * Resets button to its original state.
+         *
+         * @param {String} buttonId
+         */
+        resetButton = function(buttonId) {
+            // check if button exists
+            if (!this.items[buttonId]) {
+                return;
+            }
+
+            this.items[buttonId].initialized.then(function() {
+                resetMainListItem.call(this, this.items[buttonId].$el);
+            }.bind(this));
         },
 
         /**
@@ -445,7 +478,6 @@ define(function() {
          * @param highlight {boolean} if true a highlight effect is played
          */
         toggleEnabled = function(enabled, id, highlight) {
-
             // check if toolbar has an item with specified id
             if (!this.items[id]) {
                 return;